+++ client/api/api.go
@@ -1,4 +1,4 @@
-package client
+package api
 
 import (
 	"bytes"
