+++ client/api/api_test.go
@@ -1,4 +1,4 @@
-package client
+package api
 
 import (
 	"os"
